% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableWorkflows.R
\name{availableWorkflows}
\alias{availableWorkflows}
\title{availableWorkflows}
\usage{
availableWorkflows(quiet = F, return = F)
}
\arguments{
\item{quiet}{show available workflows}

\item{return}{return character vector of available workflows}
}
\description{
Get available workflows.
}
\examples{
availableWorkflows()
}
