% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\docType{methods}
\name{binnedData,Workflow-method}
\alias{binnedData,Workflow-method}
\title{binnedData}
\usage{
\S4method{binnedData}{Workflow}(x)
}
\arguments{
\item{x}{S4 object of class Workflow}
}
\description{
Extract binned data from a Workflow object.
}
