% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{createProject}
\alias{createProject}
\title{Create a project directory}
\usage{
createProject(
  project_name,
  workflow = availableWorkflows(),
  path = ".",
  rebuild = FALSE,
  docker = TRUE,
  github = FALSE,
  private = FALSE,
  github_actions = FALSE,
  force = FALSE
)
}
\arguments{
\item{project_name}{project name/title}

\item{workflow}{workflow to use. See \code{availableWorkflows()} for a list of available workflows.}

\item{path}{target file path for project directory}

\item{rebuild}{force rebuild of R packages installed into project \code{renv} cache}

\item{docker}{TRUE/FALSE. Create project infrastructure for building a docker container to compile the project.}

\item{github}{TRUE/FALSE. Create a GitHub repository?}

\item{private}{TRUE/FALSE. Should the GitHub repository be private? Evaluated only if argument \code{github} is TRUE.}

\item{github_actions}{TRUE/FALSE. Add Github actions infrastructure? Evaluated only if argument \code{github} is TRUE.}

\item{force}{force project creation if project directory already exists}
}
\description{
Create a project directory for a workflow.
}
\examples{
\dontrun{
createProject('test',
              workflow = 'FIE-HRMS fingerprinting',
              path = '.')
}
}
