% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{dat,Workflow-method}
\alias{dat,Workflow-method}
\title{dat}
\usage{
\S4method{dat}{Workflow}(x, type = c("raw", "pre-treated"))
}
\arguments{
\item{x}{S4 object of class Workflow}

\item{type}{get or set raw or pre-treated data}
}
\description{
return analysis data from a Workflow object.
}
