% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{defineWorkflow}
\alias{defineWorkflow}
\alias{defineWorkflow,GroverInput-method}
\alias{defineWorkflow,FilePathInput-method}
\title{Define a metabolomics workflow}
\usage{
defineWorkflow(input, workflow, project_name, ...)

\S4method{defineWorkflow}{GroverInput}(input, workflow, project_name, ...)

\S4method{defineWorkflow}{FilePathInput}(input, workflow, project_name, ...)
}
\arguments{
\item{input}{S4 object of class \code{GroverInput} or \code{FilePathInput}}

\item{workflow}{workflow name as returned by \code{availableWorkflows()}}

\item{project_name}{the project name}

\item{...}{arguments to pass to \code{defineProject}}
}
\value{
An S4 object of class \code{Workflow}
}
\description{
Define a metabolomics data analysis workflow for a given input type and workflow name.
}
\examples{
file_paths <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')
sample_information <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')

workflow_input <- filePathInput(file_paths,sample_information)

workflow_definition <- defineWorkflow(workflow_input,
                                      'FIE-HRMS fingerprinting',
                                      'Example project')
workflow_definition
}
