% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\docType{class}
\name{FilePathInput-class}
\alias{FilePathInput-class}
\title{S4 class for storing file path input information}
\description{
An S4 class for storing file path type workflow input information.
}
\section{Slots}{

\describe{
\item{\code{file_paths}}{character vector of file paths}

\item{\code{sample_information}}{a tibble containing sample information}
}}

