% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{filePathInput}
\alias{filePathInput}
\title{Workflow input using file paths}
\usage{
filePathInput(file_paths, sample_information)
}
\arguments{
\item{file_paths}{character vector of .mzML file paths}

\item{sample_information}{a tibble containing sample information. See details for specifications}
}
\value{
An S4 object of class \code{FilePathInput}.
}
\description{
Use file paths to .mzML data files for workflow input.
}
\details{
The tibble containing sample information should at least contain the following columns:
\itemize{
\item \code{fileOrder} - the numeric file order of the input files when order alphabetically as returned by list.files().
\item \code{injOrder} - the sample injection order during MS analysis
\item \code{fileName} - the sample file name
\item \code{batch} - the analytical batch number
\item \code{block} - the randomised block number
\item \code{name} - the sample name
\item \code{class} - the sample class name
}
}
\examples{
file_paths <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')
sample_information <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')

workflow_input <- filePathInput(file_paths,sample_information)

workflow_input
}
