% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/access.R
\name{files}
\alias{files}
\alias{files<-}
\alias{files,WorkflowParameters-method}
\alias{files<-,WorkflowParameters-method}
\alias{files,Workflow-method}
\title{files}
\usage{
files(x)

files(x) <- value

\S4method{files}{WorkflowParameters}(x)

\S4method{files}{WorkflowParameters}(x) <- value

\S4method{files}{Workflow}(x)
}
\arguments{
\item{x}{S4 object of class Workflow or WorflowParameters}
}
\description{
Get and set sample file paths.
}
