% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/flags.R
\docType{methods}
\name{flags}
\alias{flags}
\alias{flags,WorkflowParameters-method}
\alias{flags<-}
\title{flags}
\usage{
flags(x)

\S4method{flags}{WorkflowParameters}(x)

flags(x) <- value
}
\arguments{
\item{x}{S4 object of class WorkflowParameters}

\item{value}{new flags}
}
\description{
methods for setting and returning workflow flags
}
