% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse.R
\name{glimpse}
\alias{glimpse}
\alias{glimpse,Workflow-method}
\title{Visualise the workflow targets}
\usage{
glimpse(x)

\S4method{glimpse}{Workflow}(x)
}
\arguments{
\item{x}{S4 object of class Workflow}
}
\description{
Visualise the directed acyclic graph of a workflow targets.
}
\examples{
file_paths <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')
sample_information <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')

workflow_input <- filePathInput(file_paths,sample_information)

workflow_definition <- defineWorkflow(workflow_input,
                                      'FIE-HRMS fingerprinting',
                                      'Example project')
                 
glimpse(workflow_definition)
}
