% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{instrument}
\alias{instrument}
\alias{instrument,GroverInput-method}
\alias{instrument<-}
\alias{instrument<-,GroverInput-method}
\alias{directory}
\alias{directory,GroverInput-method}
\alias{directory<-}
\alias{directory<-,GroverInput-method}
\alias{host}
\alias{host,GroverInput-method}
\alias{host<-}
\alias{host<-,GroverInput-method}
\alias{port}
\alias{port,GroverInput-method}
\alias{port<-}
\alias{port<-,GroverInput-method}
\alias{auth}
\alias{auth,GroverInput-method}
\alias{auth<-}
\alias{auth<-,GroverInput-method}
\title{\code{GroverInput} get and set methods}
\usage{
instrument(x)

\S4method{instrument}{GroverInput}(x)

instrument(x) <- value

\S4method{instrument}{GroverInput}(x) <- value

directory(x)

\S4method{directory}{GroverInput}(x)

directory(x) <- value

\S4method{directory}{GroverInput}(x) <- value

host(x)

\S4method{host}{GroverInput}(x)

host(x) <- value

\S4method{host}{GroverInput}(x) <- value

port(x)

\S4method{port}{GroverInput}(x)

port(x) <- value

\S4method{port}{GroverInput}(x) <- value

auth(x)

\S4method{auth}{GroverInput}(x)

auth(x) <- value

\S4method{auth}{GroverInput}(x) <- value
}
\arguments{
\item{x}{S4 object of class \code{GroverInput}}

\item{value}{value to set}
}
\description{
Get and set methods for the \code{GroverInput} S4 class.
}
