% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\docType{class}
\name{GroverInput-class}
\alias{GroverInput-class}
\title{S4 class for storing grover input information}
\description{
An S4 class for storing grover type workflow input information.
}
\section{Slots}{

\describe{
\item{\code{instrument}}{analytical instrument name}

\item{\code{directory}}{experiment directory name}

\item{\code{host}}{grover API host address}

\item{\code{port}}{grover API host port}

\item{\code{auth}}{grover API host authentication code}
}}

