% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{groverInput}
\alias{groverInput}
\title{Workflow input using a grover API}
\usage{
groverInput(instrument, directory, host, port, auth)
}
\arguments{
\item{instrument}{analytical instrument name}

\item{directory}{experiment directory name}

\item{host}{grover API host address}

\item{port}{grover API host port}

\item{auth}{grover API host authentication code}
}
\value{
An S4 object of class GroverInput
}
\description{
Use .mzML files retrieved from a grover API for workflow input.
}
\details{
This specifies the retrieval of sample data files and information using a grover API. See \href{jasenfinch.github.io/grover/}{\code{grover}} for more information about mass spectral data retrieval from a grover API.
}
