% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\docType{methods}
\name{info,Workflow-method}
\alias{info,Workflow-method}
\title{info}
\usage{
\S4method{info}{Workflow}(x)

\S4method{info}{Workflow}(x)
}
\arguments{
\item{x}{S4 object of class Workflow}

\item{x}{S4 object of class Workflow}
}
\description{
Extract raw binning runinfo data from a Workflow object.

Extract raw profiling sample info from an object of class Workflow.
}
