% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/access.R
\name{info}
\alias{info}
\alias{info<-}
\alias{info,WorkflowParameters-method}
\alias{info<-,WorkflowParameters-method}
\alias{info,Workflow-method}
\title{info}
\usage{
info(x, value)

info(x) <- value

\S4method{info}{WorkflowParameters}(x)

\S4method{info}{WorkflowParameters}(x) <- value

\S4method{info}{Workflow}(x)
}
\arguments{
\item{x}{S4 object of class Workflow or WorflowParameters}

\item{value}{tibble containing sample information}
}
\description{
Get and set sample information.
}
