% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{manifest}
\alias{manifest}
\alias{manifest,Workflow-method}
\title{Workflow definition target information}
\usage{
manifest(x)

\S4method{manifest}{Workflow}(x)
}
\arguments{
\item{x}{S4 object of class Workflow}
}
\description{
Return a dataframe of information about the targets in a workflow definition.
}
\examples{
file_paths <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')
sample_information <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')

workflow_input <- inputFilePath(file_paths,sample_information)

workflow_definition <- defineWorkflow(workflow_input,
                                      'FIE-HRMS fingerprinting',
                                      'Example project')
                 
manifest(workflow_definition)
}
