% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/parameters.R
\name{parametersProcessing}
\alias{parametersProcessing}
\alias{parametersProcessing<-}
\alias{parametersProcessing,WorkflowParameters-method}
\alias{parametersProcessing<-,WorkflowParameters-method}
\title{parametersProcessing}
\usage{
parametersProcessing(parameters)

parametersProcessing(parameters) <- value

\S4method{parametersProcessing}{WorkflowParameters}(parameters)

\S4method{parametersProcessing}{WorkflowParameters}(parameters) <- value
}
\arguments{
\item{parameters}{an object of class Workflow Parameters}

\item{value}{new parameters value}
}
\description{
modify or return the processing parameters for a given WorkflowParameters object.
}
