% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{plotFeature,Workflow-method}
\alias{plotFeature,Workflow-method}
\title{plotFeature}
\usage{
\S4method{plotFeature}{Workflow}(analysis, feature, cls = "class",
  label = NULL, labelSize = 2)
}
\arguments{
\item{analysis}{object of class Workflow containing analysis results}

\item{feature}{feature to plot}

\item{cls}{info column to use for class labels}

\item{label}{info column to use for sample labels}

\item{labelSize}{sample label size}
}
\description{
Plot a pre-treated feature trend.
}
