% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{plotLDA,Workflow-method}
\alias{plotLDA,Workflow-method}
\title{plotLDA}
\usage{
\S4method{plotLDA}{Workflow}(analysis, cls = "class", label = NULL,
  scale = T, center = T, xAxis = "DF1", yAxis = "DF2",
  ellipses = T,
  title = "Principle Component - Linear Discriminant Analysis (PC-LDA)",
  legendPosition = "bottom", labelSize = 2)
}
\arguments{
\item{analysis}{object of class Workflow containing analysis results}

\item{cls}{info column to use for sample labelling}

\item{label}{info column to use for sample labels. Set to NULL for no labels.}

\item{scale}{scale the data}

\item{center}{center the data}

\item{xAxis}{principle component to plot on the x-axis}

\item{yAxis}{principle component to plot on the y-axis}

\item{ellipses}{should multivariate normal distribution 95\% confidence ellipses be plotted for each class?}

\item{title}{plot title}

\item{legendPosition}{legend position to pass to legend.position argument of \code{ggplot2::theme}}

\item{labelSize}{label size. Ignored if \code{label} is \code{NULL}}
}
\description{
Plot linear discriminant analysis resultus of pre-treated data
}
