% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{plotPCA,Workflow-method}
\alias{plotPCA,Workflow-method}
\title{plotPCA}
\usage{
\S4method{plotPCA}{Workflow}(analysis, cls = "class", label = NULL,
  scale = T, center = T, xAxis = "PC1", yAxis = "PC2",
  ellipses = T, title = "Principle Component Analysis (PCA)",
  legendPosition = "bottom", labelSize = 2)
}
\arguments{
\item{analysis}{object of class Workflow containing analysis results}

\item{cls}{info column to use for sample labelling}

\item{label}{info column to use for sample labels. Set to NULL for no labels.}

\item{scale}{scale the data}

\item{center}{center the data}

\item{xAxis}{principle component to plot on the x-axis}

\item{yAxis}{principle component to plot on the y-axis}

\item{ellipses}{should multivariate normal distribution 95\% confidence ellipses be plotted for each class?}

\item{title}{plot title}

\item{legendPosition}{legend position to pass to legend.position argument of \code{ggplot2::theme}}

\item{labelSize}{label size. Ignored if \code{label} is \code{NULL}}
}
\description{
Plot principle component analysis results of pre-treated data.
}
