% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotRSD,Workflow-method}
\alias{plotRSD,Workflow-method}
\title{plotRSD}
\usage{
\S4method{plotRSD}{Workflow}(
  analysis,
  cls = "class",
  QCidx = "QC",
  QCparameters = NULL,
  histBins = 30,
  title = "Relative standard deviation distributions"
)
}
\arguments{
\item{analysis}{object of class Workflow containing analysis results}

\item{cls}{info column to use for class labels}

\item{QCidx}{QC sample label}

\item{QCparameters}{alternative parameters for QC sample pre-treatment. See details}

\item{histBins}{number of bins to use for histogram plotting}

\item{title}{plot title}
}
\description{
Plot RSD distributions of raw data in quality control samples.
}
\details{
If QCparameters is set as \code{NULL}, the default QC pre-treatment parameters are used as given by \code{analysisParameters('preTreat')}. Alternative pre-treatment routines can be used by specifying an \code{AnalysisParameters} object for \code{QCparameters}.
}
