% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotting.R
\docType{methods}
\name{plotTIC}
\alias{plotTIC}
\alias{plotTIC,Workflow-method}
\title{plotTIC}
\usage{
plotTIC(x, ...)

\S4method{plotTIC}{Workflow}(x, ...)
}
\arguments{
\item{x}{S4 object of class Workflow}

\item{...}{arguments to pass to individal object methods depending on what spectral processing has been used (see \code{?binneR::plotTIC} or \code{?profilePro::plotTIC})}
}
\description{
Plot processed sample total ion counts.
}
