% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{preTreated,Workflow-method}
\alias{preTreated,Workflow-method}
\title{preTreated}
\usage{
\S4method{preTreated}{Workflow}(x)
}
\arguments{
\item{x}{S4 object of class Workflow}
}
\description{
return pre-treated AnalysisData object from a Workflow object.
}
