% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\docType{methods}
\name{preTreatedData,Workflow-method}
\alias{preTreatedData,Workflow-method}
\title{preTreatedData}
\usage{
\S4method{preTreatedData}{Workflow}(x)
}
\arguments{
\item{x}{S4 object of class Workflow}
}
\description{
return pre-treated data from a Workflow object.
}
