% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\docType{methods}
\name{preTreatedInfo,Workflow-method}
\alias{preTreatedInfo,Workflow-method}
\title{preTreatedData}
\usage{
\S4method{preTreatedInfo}{Workflow}(x)
}
\arguments{
\item{x}{S4 object of class Workflow}
}
\description{
return pre-treated sample information from a Workflow object.
}
