% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{processedData,Workflow-method}
\alias{processedData,Workflow-method}
\title{processedData}
\usage{
\S4method{processedData}{Workflow}(x)
}
\arguments{
\item{x}{S4 object of class Workflow}
}
\description{
Extract processed metabolomics data from an object of class Workflow
}
