% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/access.R
\docType{methods}
\name{processedInfo}
\alias{processedInfo}
\alias{processedInfo,Workflow-method}
\title{processedInfo}
\usage{
processedInfo(x)

\S4method{processedInfo}{Workflow}(x)
}
\arguments{
\item{x}{S4 object of class Workflow}
}
\description{
Extract processed sample info from an object of class Workflow.
}
