% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{projectName}
\alias{projectName}
\alias{projectName,Project-method}
\alias{projectName<-}
\alias{projectName<-,Project-method}
\alias{path}
\alias{path,Project-method}
\alias{path<-}
\alias{path<-,Project-method}
\alias{renv}
\alias{renv,Project-method}
\alias{renv<-}
\alias{renv<-,Project-method}
\alias{docker}
\alias{docker,Project-method}
\alias{docker<-}
\alias{docker<-,Project-method}
\alias{github}
\alias{github,Project-method}
\alias{github<-}
\alias{github<-,Project-method}
\alias{private}
\alias{private,Project-method}
\alias{private<-}
\alias{private<-,Project-method}
\alias{githubActions}
\alias{githubActions,Project-method}
\alias{githubActions<-}
\alias{githubActions<-,Project-method}
\alias{force}
\alias{force,Project-method}
\alias{force<-}
\alias{force<-,Project-method}
\title{\code{Project} class get and set methods}
\usage{
projectName(x)

\S4method{projectName}{Project}(x)

projectName(x) <- value

\S4method{projectName}{Project}(x) <- value

path(x)

\S4method{path}{Project}(x)

path(x) <- value

\S4method{path}{Project}(x) <- value

renv(x)

\S4method{renv}{Project}(x)

renv(x) <- value

\S4method{renv}{Project}(x) <- value

docker(x)

\S4method{docker}{Project}(x)

docker(x) <- value

\S4method{docker}{Project}(x) <- value

github(x)

\S4method{github}{Project}(x)

github(x) <- value

\S4method{github}{Project}(x) <- value

private(x)

\S4method{private}{Project}(x)

private(x) <- value

\S4method{private}{Project}(x) <- value

githubActions(x)

\S4method{githubActions}{Project}(x)

githubActions(x) <- value

\S4method{githubActions}{Project}(x) <- value

force(x)

\S4method{force}{Project}(x)

force(x) <- value

\S4method{force}{Project}(x) <- value
}
\arguments{
\item{x}{S4 object of class \code{Project}}

\item{value}{value to set}
}
\description{
Get and set methods for the \code{Project} S4 class.
}
\examples{
workflow_project <- defineProject('A metabolomics project')

## Return the project name
projectName(workflow_project)

## Set the project name
projectName(workflow_project) <- 'A new name'

## Return the project directory path
path(workflow_project)

## Set the project directory path
path(workflow_project) <- './a_directory'

## Return the project renv option
renv(workflow_project)

## Set the project renv option
renv(workflow_project) <- FALSE

## Return the project docker option
docker(workflow_project)

## Set the project docker option
docker(workflow_project) <- FALSE

## Return the project github option
github(workflow_project)

## Set the project github option
github(workflow_project) <- TRUE

## Return the project private option
private(workflow_project)

## Set the project private option
private(workflow_project) <- TRUE

## Return the project github actions option
githubActions(workflow_project)

## Set the project github actions option
githubActions(workflow_project) <- TRUE

## Return the project force option
force(workflow_project)

## Set the project force option
force(workflow_project) <- TRUE
}
