% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\docType{class}
\name{Project-class}
\alias{Project-class}
\title{S4 class to store project directory definitions}
\description{
An S4 class to store the workflow project directory definitions.
}
\section{Slots}{

\describe{
\item{\code{project_name}}{the project name}

\item{\code{path}}{the project directory path}

\item{\code{renv}}{add infrastructure for reproducible R package environment management from the \code{renv} package}

\item{\code{docker}}{creation for project docker infrastructure}

\item{\code{github}}{creation of a GitHub repository}

\item{\code{private}}{private GitHub repository}

\item{\code{github_actions}}{creation of GitHub actions infrastructure}

\item{\code{force}}{force project creation if the project directory already exists}
}}

