% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\docType{methods}
\name{resultsAnalysis,Workflow-method}
\alias{resultsAnalysis,Workflow-method}
\title{resultsAnalysis}
\usage{
\S4method{resultsAnalysis}{Workflow}(object)
}
\arguments{
\item{object}{S4 object of class Workflow}
}
\description{
return analysis results from a Workflow object
}
