% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/results.R
\docType{methods}
\name{resultsAnalysis}
\alias{resultsAnalysis}
\alias{resultsAnalysis,Workflow-method}
\title{@rdname resultsAnalysis}
\usage{
resultsAnalysis(object)

\S4method{resultsAnalysis}{Workflow}(object)
}
\arguments{
\item{object}{S4 object of class Workflow}
}
\description{
@rdname resultsAnalysis

return analysis results from a Workflow object
}
