% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\docType{methods}
\name{resultsAnnotation,Workflow-method}
\alias{resultsAnnotation,Workflow-method}
\title{resultsAnnotation}
\usage{
\S4method{resultsAnnotation}{Workflow}(object)
}
\arguments{
\item{object}{S4 object of class Workflow}
}
\description{
return annotation results from a Workflow object
}
