% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/results.R
\docType{methods}
\name{resultsAnnotation}
\alias{resultsAnnotation}
\alias{resultsAnnotation,Workflow-method}
\title{@rdname resultsAnnotation}
\usage{
resultsAnnotation(object)

\S4method{resultsAnnotation}{Workflow}(object)
}
\arguments{
\item{object}{S4 object of class Workflow}
}
\description{
@rdname resultsAnnotation

return annotation results from a Workflow object
}
