% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/results.R
\docType{methods}
\name{resultsParameters}
\alias{resultsParameters}
\alias{resultsParameters,Workflow-method}
\title{resultsParameters}
\usage{
resultsParameters(object)

\S4method{resultsParameters}{Workflow}(object)
}
\arguments{
\item{object}{S4 object of class Workflow}
}
\description{
return parameters from a Workflow object
}
