% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/results.R
\docType{methods}
\name{resultsProcessing}
\alias{resultsProcessing}
\alias{resultsProcessing,Workflow-method}
\title{resultsProcessing}
\usage{
resultsProcessing(object)

\S4method{resultsProcessing}{Workflow}(object)
}
\arguments{
\item{object}{S4 object of class Workflow}
}
\description{
return processing results from a Workflow object
}
