% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retryWorkflow.R
\name{retryWorkflow}
\alias{retryWorkflow}
\title{retryWorkflow}
\usage{
retryWorkflow(analysis)
}
\arguments{
\item{analysis}{an S4 object of class Workflow}
}
\description{
Retry a workflow after initial analysis with \code{workflow} following an error. 
Will resume the analysis from the point of error.
}
