% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-method.R
\name{show,Workflow-method}
\alias{show,Workflow-method}
\title{show-Workflow}
\usage{
\S4method{show}{Workflow}(object)
}
\arguments{
\item{object}{S4 object of class Workflow}
}
\description{
show method for Workflows class
}
