% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{sinfo,Workflow-method}
\alias{sinfo,Workflow-method}
\title{sinfo}
\usage{
\S4method{sinfo}{Workflow}(x, type = c("raw", "pre-treated"))
}
\arguments{
\item{x}{S4 object of class Workflow}

\item{type}{get or set raw or pre-treated data}
}
\description{
return sample information from a Workflow object.
}
