% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{name}
\alias{name}
\alias{name,Target-method}
\alias{name<-}
\alias{name<-,Target-method}
\alias{command}
\alias{command,Target-method}
\alias{command<-}
\alias{command<-,Target-method}
\alias{type}
\alias{type,Target-method}
\alias{type<-}
\alias{type<-,Target-method}
\alias{pattern}
\alias{pattern,Target-method}
\alias{pattern<-}
\alias{pattern<-,Target-method}
\alias{args}
\alias{args,Target-method}
\alias{args<-}
\alias{args<-,Target-method}
\alias{code}
\alias{code,Target-method}
\title{\code{Target} class get and set methods}
\usage{
name(x)

\S4method{name}{Target}(x)

name(x) <- value

\S4method{name}{Target}(x) <- value

command(x)

\S4method{command}{Target}(x)

command(x) <- value

\S4method{command}{Target}(x) <- value

type(x)

\S4method{type}{Target}(x)

type(x) <- value

\S4method{type}{Target}(x) <- value

pattern(x)

\S4method{pattern}{Target}(x)

pattern(x) <- value

\S4method{pattern}{Target}(x) <- value

args(x)

\S4method{args}{Target}(x)

args(x) <- value

\S4method{args}{Target}(x) <- value

code(x)

\S4method{code}{Target}(x)
}
\arguments{
\item{x}{S4 object of class Target}

\item{value}{value to set}
}
\description{
Get and set methods for the Target S4 class.
}
