% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\docType{class}
\name{Target-class}
\alias{Target-class}
\title{S4 class to store target a declaration}
\description{
An S4 class for storing a workflow target declaration.
}
\section{Slots}{

\describe{
\item{\code{name}}{the target name}

\item{\code{command}}{the R code to run the target}

\item{\code{type}}{the target archetype}

\item{\code{args}}{arguments to pass to the specified target archetype}
}}

