% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{target}
\alias{target}
\title{Create a workflow target definition}
\usage{
target(name, command, type = "tar_target", args = list())
}
\arguments{
\item{name}{the target name}

\item{command}{the R command to run the target}

\item{type}{the target archetype}

\item{args}{a list of arguments to pass the the specified target archetype}
}
\value{
An S4 object of class Target.
}
\description{
Create a workflow target definition.
}
\details{
Target types can be one of any provided by the \code{targets} or \code{tarchetypes} packages.
}
\examples{
workflow_target <- target('a_target',1 + 1,args = list(memory = 'persistent'))

workflow_target
}
