% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{targetsWorkflow}
\alias{targetsWorkflow}
\alias{targetsWorkflow,Workflow-method}
\alias{targetsInput}
\alias{targetsInput,FilePathInput-method}
\alias{targetsInput,GroverInput-method}
\alias{targetsInput,Workflow-method}
\alias{targetsSpectralProcessing}
\alias{targetsPretreatment}
\alias{targetsMFassignment}
\alias{targetsModelling}
\alias{targetsCorrelations}
\alias{targetsReport}
\title{Workflow pipeline targets}
\usage{
targetsWorkflow(x)

\S4method{targetsWorkflow}{Workflow}(x)

targetsInput(x)

\S4method{targetsInput}{FilePathInput}(x)

\S4method{targetsInput}{GroverInput}(x)

\S4method{targetsInput}{Workflow}(x)

targetsSpectralProcessing(x)

targetsPretreatment(x)

targetsMFassignment(x)

targetsModelling(x)

targetsCorrelations(x)

targetsReport(x)
}
\arguments{
\item{x}{the workflow type or an S4 object of class \code{Workflow}, \code{FilePathInput} or \code{GroverInput}}
}
\value{
A list of \code{Target} S4 class target definitions.
}
\description{
Target definitions for workflow input.
}
\examples{
## Full workflow example
file_paths <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')
sample_information <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')

workflow_input <- inputFilePath(file_paths,sample_information)

workflow_definition <- defineWorkflow(workflow_input,
                                     'FIE-HRMS fingerprinting',
                                     'Example project')

targetsWorkflow(workflow_definition)

## Examples for individual modules
targetsSpectralProcessing('FIE-HRMS fingerprinting')
targetsInput(workflow_input)
targetsPretreatment('FIE-HRMS fingerprinting')
targetsMFassignment('FIE-HRMS fingerprinting')
targetsModelling('FIE-HRMS fingerprinting')
targetsCorrelations('FIE-HRMS fingerprinting')
}
