% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\alias{validate,Workflow-method}
\title{Validate a workflow definition}
\usage{
validate(x)

\S4method{validate}{Workflow}(x)
}
\arguments{
\item{x}{S4 object of class Workflow}
}
\description{
Validate a workflow definition, checking for issues.
An error or a warning will be returned if a warning is detected.
}
\examples{
file_paths <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')
sample_information <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')

workflow_input <- inputFilePath(file_paths,sample_information)

workflow_definition <- defineWorkflow(workflow_input,
                                      'FIE-HRMS fingerprinting',
                                      'Example project')
                 
validate(workflow_definition)
}
