% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{Workflow-class}
\alias{Workflow-class}
\title{Workflow}
\description{
An S4 class to store workflow results
}
\section{Slots}{

\describe{
\item{\code{logs}}{list containing workflow dates and times}

\item{\code{flags}}{character containing workflow sections completed}

\item{\code{files}}{list file names to process}

\item{\code{workflowParameters}}{WorkflowParameters object containing parameters used}

\item{\code{processed}}{Processed object containing processing results}

\item{\code{analysed}}{Analysis object containing analysis results}

\item{\code{annotated}}{Assignment object containing annotation results}
}}

\seealso{
\code{\link[metabolyseR]{Analysis-class}} \code{\link[MFassign]{Assignment-class}} \code{\link[binneRlyse]{Binalysis-class}} \code{\link[profilePro]{MetaboProfile-class}}
}
