% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit.R
\name{moduleReplace}
\alias{moduleReplace}
\alias{moduleReplace,Workflow-method}
\alias{moduleRemove}
\alias{moduleRemove,Workflow-method}
\alias{moduleAdd}
\alias{moduleAdd,Workflow-method}
\alias{targetReplace}
\alias{targetReplace,Workflow-method}
\alias{targetRemove}
\alias{targetRemove,Workflow-method}
\alias{targetAdd}
\alias{targetAdd,Workflow-method}
\title{Edit workflow targets and modules}
\usage{
moduleReplace(x, module, replacement)

\S4method{moduleReplace}{Workflow}(x, module, replacement)

moduleRemove(x, module)

\S4method{moduleRemove}{Workflow}(x, module)

moduleAdd(x, module, addition)

\S4method{moduleAdd}{Workflow}(x, module, addition)

targetReplace(x, module, target, replacement)

\S4method{targetReplace}{Workflow}(x, module, target, replacement)

targetRemove(x, module, target)

\S4method{targetRemove}{Workflow}(x, module, target)

targetAdd(x, module, target, addition)

\S4method{targetAdd}{Workflow}(x, module, target, addition)
}
\arguments{
\item{x}{S4 object of class \code{Workflow}}

\item{module}{the workflow module name}

\item{replacement}{the replacement workflow module or target}

\item{addition}{the workflow module or target to add}

\item{target}{the workflow target name}
}
\value{
S4 object of class \code{Workflow} with edited targets
}
\description{
Edit the targets of a workflow definition.
}
\examples{
## Define a workflow
file_paths <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')
sample_information <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')

workflow_input <- filePathInput(file_paths,sample_information)

workflow_definition <- defineWorkflow(workflow_input,
                                     'FIE-HRMS fingerprinting',
                                     'Example project')
                                     
## Add a module
workflow_definition <- moduleAdd(workflow_definition,
                                 'additional_module',
                                 list(new_target = target('new_target','1 + 1')))
## Replace the additional module
workflow_definition <- moduleReplace(workflow_definition,
                                     'additional_module',
                                     list(replacement_target = target('replacement_target',
                                     '1 + 2')))
## Remove the additional module
workflow_definition <- moduleRemove(workflow_definition,
                                    'additional_module')

## Add a target to the input module
workflow_definition <- targetAdd(workflow_definition,
                                 'input',
                                 'new_target',
                                 target('new_target','1 + 1'))
## Replace the additional target
workflow_definition <- targetReplace(workflow_definition,
                                 'input',
                                 'new_target',
                                 target('new_target','1 + 2')) 
## Remove the additional target
workflow_definition <- targetRemove(workflow_definition,
                                 'input',
                                 'new_target')
}
