% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{workflowFlags}
\alias{workflowFlags}
\title{workflowFlags}
\usage{
workflowFlags(w)
}
\arguments{
\item{w}{workflow name as returned by \code{workflowParameters()}}
}
\description{
Return possible flags for a given workflow.
}
\examples{
workflowFlags('FIE-HRMS fingerprinting')
}
