% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflowParameters.R
\name{workflowParameters}
\alias{workflowParameters}
\title{workflowParameters}
\usage{
workflowParameters(workflow, files, info, cls = "class", ...)
}
\arguments{
\item{workflow}{the workflow analysis to use. NULL prints the available workflows.}

\item{files}{supply vector of file paths for auto detection of parameters. Currently only enabled for the FIE_HRMSfingerprinting workflow}

\item{...}{arguments to pass to binneR::detectParameters}
}
\description{
Initiate default workflow parameters for a selected workflow.
}
\examples{
files <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes') 
info <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')

wp <- workflowParameters('FIE-HRMS fingerprinting',files,info)
}
