% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{workflowTargets}
\alias{workflowTargets}
\alias{workflowTargets,Workflow-method}
\alias{inputTargets}
\alias{inputTargets,FilePathInput-method}
\alias{inputTargets,GroverInput-method}
\alias{inputTargets,Workflow-method}
\alias{spectralProcessingTargets}
\alias{spectralProcessingTargets,Workflow-method}
\alias{pretreatmentTargets}
\alias{pretreatmentTargets,Workflow-method}
\title{Workflow input targets}
\usage{
workflowTargets(x)

\S4method{workflowTargets}{Workflow}(x)

inputTargets(x)

\S4method{inputTargets}{FilePathInput}(x)

\S4method{inputTargets}{GroverInput}(x)

\S4method{inputTargets}{Workflow}(x)

spectralProcessingTargets(x)

\S4method{spectralProcessingTargets}{Workflow}(x)

pretreatmentTargets(x)

\S4method{pretreatmentTargets}{Workflow}(x)
}
\arguments{
\item{x}{S4 object of class \code{FilePathInput} or \code{GroverInput}}
}
\value{
A list of \code{Target} S4 class target definitions.
}
\description{
Target definitions for workflow input.
}
\examples{
file_paths <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')
sample_information <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')

workflow_input <- filePathInput(file_paths,sample_information)

inputTargets(workflow_input)
}
