% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{availableWorkflows}
\alias{availableWorkflows}
\title{Available workflows}
\usage{
availableWorkflows()
}
\value{
A character vector of available workflow names.
}
\description{
Return the available metabolomics workflows.
}
\examples{
availableWorkflows()
}
