% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{createProject}
\alias{createProject}
\alias{createProject,Project-method}
\title{Create a project directory}
\usage{
createProject(project, workflow)

\S4method{createProject}{Project}(project, workflow)
}
\arguments{
\item{project}{S4 object of class \code{Project}}

\item{workflow}{workflow to use. See \code{availableWorkflows()} for a list of available workflows.}
}
\description{
Create a project directory for a workflow.
}
\examples{
\dontrun{
workflow_project <- defineProject('A metabolomics project')

createProject(workflow_project,
              workflow = 'FIE-HRMS fingerprinting')
}
}
