% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{defineProject}
\alias{defineProject}
\title{Define a project directory}
\usage{
defineProject(
  project_name,
  path = ".",
  renv = TRUE,
  docker = TRUE,
  github = FALSE,
  private = FALSE,
  github_actions = FALSE,
  force = FALSE
)
}
\arguments{
\item{project_name}{project name/title}

\item{path}{target file path for project directory}

\item{renv}{add infrastructure for reproducible R package environment management from the \code{renv} package}

\item{docker}{TRUE/FALSE. Create project infrastructure for building a docker container to compile the project.}

\item{github}{TRUE/FALSE. Create a GitHub repository?}

\item{private}{TRUE/FALSE. Should the GitHub repository be private? Evaluated only if argument \code{github} is TRUE.}

\item{github_actions}{TRUE/FALSE. Add Github actions infrastructure? Evaluated only if argument \code{github} is TRUE.}

\item{force}{force project creation if project directory already exists}
}
\value{
An S4 object of class \code{Project}.
}
\description{
Define the project directory for a workflow
}
\examples{
workflow_project <- defineProject('A metabolomics project')
}
