% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{instrument}
\alias{instrument}
\alias{instrument,GroverInput-method}
\alias{instrument<-}
\alias{instrument<-,GroverInput-method}
\alias{directory}
\alias{directory,GroverInput-method}
\alias{directory<-}
\alias{directory<-,GroverInput-method}
\alias{host}
\alias{host,GroverInput-method}
\alias{host<-}
\alias{host<-,GroverInput-method}
\alias{port}
\alias{port,GroverInput-method}
\alias{port<-}
\alias{port<-,GroverInput-method}
\alias{auth}
\alias{auth,GroverInput-method}
\alias{auth<-}
\alias{auth<-,GroverInput-method}
\title{\code{GroverInput} get and set methods}
\usage{
instrument(x)

\S4method{instrument}{GroverInput}(x)

instrument(x) <- value

\S4method{instrument}{GroverInput}(x) <- value

directory(x)

\S4method{directory}{GroverInput}(x)

directory(x) <- value

\S4method{directory}{GroverInput}(x) <- value

host(x)

\S4method{host}{GroverInput}(x)

host(x) <- value

\S4method{host}{GroverInput}(x) <- value

port(x)

\S4method{port}{GroverInput}(x)

port(x) <- value

\S4method{port}{GroverInput}(x) <- value

auth(x)

\S4method{auth}{GroverInput}(x)

auth(x) <- value

\S4method{auth}{GroverInput}(x) <- value
}
\arguments{
\item{x}{S4 object of class \code{GroverInput}}

\item{value}{value to set}
}
\description{
Get and set methods for the \code{GroverInput} S4 class.
}
\examples{
workflow_input <- inputGrover(instrument = 'An_instrument',
                              directory = 'Experiment_directory',
                              host = 'a.grover.host',
                              port = 80,
                              auth = '1234')

## Return input instrument
instrument(workflow_input)

## Set the input instrument
instrument(workflow_input) <- 'A_different_instrument'

## Return input directory
directory(workflow_input)

## Set the input directory
directory(workflow_input) <- 'Another_experiment'

## Return input host
host(workflow_input)

## Set the input host
host(workflow_input) <- 'a.new.host'

## Return the input port
port(workflow_input)

## Set the input port
port(workflow_input) <- 81

## Return the input auth
auth(workflow_input)

## Set the input auth
auth(workflow_input) <- 'abcd'
}
