% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametersAnalysis.R
\docType{methods}
\name{parametersAnalysis,WorkflowParameters-method}
\alias{parametersAnalysis,WorkflowParameters-method}
\alias{parametersAnalysis<-}
\title{parametersAnalysis}
\usage{
\S4method{parametersAnalysis}{WorkflowParameters}(parameters)

parametersAnalysis(parameters) <- value
}
\arguments{
\item{parameters}{an object of class Workflow Parameters}
}
\description{
modify or return the Analysis parameters for a given WorkflowParameters object.
}
