% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/parametersAnalysis.R
\name{parametersAnalysis}
\alias{parametersAnalysis}
\alias{parametersAnalysis,WorkflowParameters-method}
\alias{parametersAnalysis<-}
\title{parametersAnalysis}
\usage{
parametersAnalysis(parameters)

\S4method{parametersAnalysis}{WorkflowParameters}(parameters)

parametersAnalysis(parameters) <- value
}
\arguments{
\item{parameters}{an object of class Workflow Parameters}

\item{value}{new parameters value}
}
\description{
modify or return the Analysis parameters for a given WorkflowParameters object.
}
