% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/parameters.R
\name{parametersAnnotation}
\alias{parametersAnnotation}
\alias{parametersAnnotation<-}
\alias{parametersAnnotation,WorkflowParameters-method}
\alias{parametersAnnotation<-,WorkflowParameters-method}
\title{parametersAnnotation}
\usage{
parametersAnnotation(parameters)

parametersAnnotation(parameters) <- value

\S4method{parametersAnnotation}{WorkflowParameters}(parameters)

\S4method{parametersAnnotation}{WorkflowParameters}(parameters) <- value
}
\arguments{
\item{parameters}{an object of class Workflow Parameters}

\item{value}{new parameters value}
}
\description{
modify or return the Annotation parameters for a given WorkflowParameters object.
}
