% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametersProcessing.R
\docType{methods}
\name{parametersProcessing,WorkflowParameters-method}
\alias{parametersProcessing,WorkflowParameters-method}
\alias{parametersProcessing<-}
\title{parametersProcessing}
\usage{
\S4method{parametersProcessing}{WorkflowParameters}(parameters)

parametersProcessing(parameters) <- value
}
\arguments{
\item{parameters}{an object of class Workflow Parameters}
}
\description{
modify or return the processing parameters for a given WorkflowParameters object.
}
