% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{name}
\alias{name}
\alias{name,Target-method}
\alias{name<-}
\alias{name<-,Target-method}
\alias{command}
\alias{command,Target-method}
\alias{command<-}
\alias{command<-,Target-method}
\alias{type}
\alias{type,Target-method}
\alias{type<-}
\alias{type<-,Target-method}
\alias{args}
\alias{args,Target-method}
\alias{args<-}
\alias{args<-,Target-method}
\alias{comment}
\alias{comment,Target-method}
\alias{comment<-}
\alias{comment<-,Target-method}
\alias{code}
\alias{code,Target-method}
\title{\code{Target} class get and set methods}
\usage{
name(x)

\S4method{name}{Target}(x)

name(x) <- value

\S4method{name}{Target}(x) <- value

command(x)

\S4method{command}{Target}(x)

command(x) <- value

\S4method{command}{Target}(x) <- value

type(x)

\S4method{type}{Target}(x)

type(x) <- value

\S4method{type}{Target}(x) <- value

args(x)

\S4method{args}{Target}(x)

args(x) <- value

\S4method{args}{Target}(x) <- value

comment(x)

\S4method{comment}{Target}(x)

comment(x) <- value

\S4method{comment}{Target}(x) <- value

code(x)

\S4method{code}{Target}(x)
}
\arguments{
\item{x}{S4 object of class Target}

\item{value}{value to set}
}
\description{
Get and set methods for the Target S4 class.
}
\examples{
workflow_target <- target('a_target',
                          1 + 1,
                          args = list(memory = 'persistent'), 
                          comment = 'A target')

## Return the target name
name(workflow_target)

## Set the target name
name(workflow_target) <- 'a_new_name'

## Return the target R code
command(workflow_target)

## Set the target R code
command(workflow_target) <- rlang::expr(1 * 2)

## Return the target type
type(workflow_target)

## Set the target type
type(workflow_target) <- 'tarchetypes::tar_file'

## Return the list target arguments
args(workflow_target)

## Set the target arguments
args(workflow_target) <- list(error = 'continue')

## Return the target comment
comment(workflow_target)

## Set the workflow comment
comment(workflow_target) <- 'A new comment'
}
