% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\docType{class}
\name{Target-class}
\alias{Target-class}
\title{S4 class to store a target definition}
\description{
An S4 class for storing a workflow target definition.
}
\section{Slots}{

\describe{
\item{\code{name}}{the target name}

\item{\code{command}}{the R code to run the target as a string}

\item{\code{type}}{the target archetype}

\item{\code{args}}{arguments to pass to the specified target archetype}
}}

