% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{target}
\alias{target}
\title{Create a workflow target definition}
\usage{
target(
  name,
  command,
  type = "tar_target",
  args = list(),
  comment = character()
)
}
\arguments{
\item{name}{the target name}

\item{command}{the R command to run the target}

\item{type}{the target archetype}

\item{args}{a list of arguments to pass the the specified target archetype}

\item{comment}{optional comment that precedes the target code}
}
\value{
An S4 object of class Target.
}
\description{
Create a workflow target definition.
}
\details{
The specified target \code{type} can be one of any provided by the \code{targets} package such as \code{tar_target}.
For target archetypes outside of \code{targets}, declare the package directly for argument
\code{type}. For instance, \code{type = "tarchetypes::tar_file"}.
}
\examples{
workflow_target <- target('a_target',
                          1 + 1,
                          args = list(memory = 'persistent'), 
                          comment = 'A target')

workflow_target
}
