% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{targetsWorkflow}
\alias{targetsWorkflow}
\alias{targetsWorkflow,Workflow-method}
\alias{targetsInput}
\alias{targetsInput,FilePathInput-method}
\alias{targetsInput,GroverInput-method}
\alias{targetsInput,Workflow-method}
\alias{targetsSpectralProcessing}
\alias{targetsSpectralProcessing,Workflow-method}
\alias{targetsPretreatment}
\alias{targetsPretreatment,Workflow-method}
\alias{targetsMFassignment}
\alias{targetsMFassignment,Workflow-method}
\alias{targetsModelling}
\alias{targetsModelling,Workflow-method}
\alias{targetsCorrelations}
\alias{targetsCorrelations,Workflow-method}
\title{Workflow pipeline targets}
\usage{
targetsWorkflow(x)

\S4method{targetsWorkflow}{Workflow}(x)

targetsInput(x)

\S4method{targetsInput}{FilePathInput}(x)

\S4method{targetsInput}{GroverInput}(x)

\S4method{targetsInput}{Workflow}(x)

targetsSpectralProcessing(x)

\S4method{targetsSpectralProcessing}{Workflow}(x)

targetsPretreatment(x)

\S4method{targetsPretreatment}{Workflow}(x)

targetsMFassignment(x)

\S4method{targetsMFassignment}{Workflow}(x)

targetsModelling(x)

\S4method{targetsModelling}{Workflow}(x)

targetsCorrelations(x)

\S4method{targetsCorrelations}{Workflow}(x)
}
\arguments{
\item{x}{S4 object of class \code{FilePathInput} or \code{GroverInput}}
}
\value{
A list of \code{Target} S4 class target definitions.
}
\description{
Target definitions for workflow input.
}
\examples{
file_paths <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')
sample_information <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')

workflow_input <- filePathInput(file_paths,sample_information)

workflow_definition <- defineWorkflow(workflow_input,
                                     'FIE-HRMS fingerprinting',
                                     'Example project')

targetsWorkflow(workflow_input)
}
