% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{type,Workflow-method}
\alias{type,Workflow-method}
\alias{type<-,Workflow-method}
\alias{input}
\alias{input,Workflow-method}
\alias{input<-}
\alias{input<-,Workflow-method}
\alias{targets}
\alias{targets,Workflow-method}
\alias{targets<-}
\alias{targets<-,Workflow-method}
\alias{modules}
\alias{modules,Workflow-method}
\alias{filePaths,Workflow-method}
\alias{filePaths<-,Workflow-method}
\alias{sampleInformation,Workflow-method}
\alias{sampleInformation<-,Workflow-method}
\alias{instrument,Workflow-method}
\alias{instrument<-,Workflow-method}
\alias{directory,Workflow-method}
\alias{directory<-,Workflow-method}
\alias{host,Workflow-method}
\alias{host<-,Workflow-method}
\alias{port,Workflow-method}
\alias{port<-,Workflow-method}
\alias{auth,Workflow-method}
\alias{auth<-,Workflow-method}
\title{\code{Workflow} class get and set methods}
\usage{
\S4method{type}{Workflow}(x)

\S4method{type}{Workflow}(x) <- value

input(x)

\S4method{input}{Workflow}(x)

input(x) <- value

\S4method{input}{Workflow}(x) <- value

targets(x)

\S4method{targets}{Workflow}(x)

targets(x) <- value

\S4method{targets}{Workflow}(x) <- value

modules(x)

\S4method{modules}{Workflow}(x)

\S4method{filePaths}{Workflow}(x)

\S4method{filePaths}{Workflow}(x) <- value

\S4method{sampleInformation}{Workflow}(x)

\S4method{sampleInformation}{Workflow}(x) <- value

\S4method{instrument}{Workflow}(x)

\S4method{instrument}{Workflow}(x) <- value

\S4method{directory}{Workflow}(x)

\S4method{directory}{Workflow}(x) <- value

\S4method{host}{Workflow}(x)

\S4method{host}{Workflow}(x) <- value

\S4method{port}{Workflow}(x)

\S4method{port}{Workflow}(x) <- value

\S4method{auth}{Workflow}(x)

\S4method{auth}{Workflow}(x) <- value
}
\arguments{
\item{x}{S4 object of class \code{Workflow}}

\item{value}{value to set}
}
\description{
Get and set methods for the \code{Workflow} S4 class.
}
