% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/workflow.R
\name{workflow}
\alias{workflow}
\alias{workflow,WorkflowParameters-method}
\title{workflow}
\usage{
workflow(parameters)

\S4method{workflow}{WorkflowParameters}(parameters)
}
\arguments{
\item{parameters}{S4 object of class WorkflowParameters containing the workflow parameters.}
}
\description{
execute workflow based on specified worklow parameters and a file list.
}
\examples{
\dontrun{

files <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes') 
info <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')

wp <- workflowParameters('FIE-HRMS fingerprinting',files,info)
analysis <- workflow(wp)
}
}
