% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow}
\alias{workflow}
\title{workflow}
\usage{
workflow(files, parameters)
}
\description{
workflow
}
\examples{
\dontrun{
analysis <- workflow(
   list.files(
       system.file(
           'DataSets/FIE-HRMS/BdistachyonEcotypes',
           package = 'metaboData'),
       full.names = TRUE), 
       workflowParameters('FIE')
   )}
}
