% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow}
\alias{workflow}
\title{workflow}
\usage{
workflow(files, info, parameters)
}
\arguments{
\item{files}{list of file paths to raw data}

\item{info}{tibble containing sample information}

\item{parameters}{S4 object of class WorkflowParameters containing the workflow parameters.}
}
\description{
execute workflow based on specified worklow parameters and a file list.
}
\examples{
\dontrun{
library(magrittr)
files <- list.files(
       system.file(
           'DataSets/FIE-HRMS/BdistachyonEcotypes',
           package = 'metaboData'),
       full.names = TRUE)
info <- readr::read_csv(files[grepl('runinfo',files)]) \%>\%
       arrange(fileName) \%>\% 
       mutate(fileOrder = 1:nrow(.))
files <- list(files[!grepl('runinfo',files)])

analysis <- workflow(files, info, workflowParameters('FIE_HRMSfingerprinting'))
}
}
