% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow}
\alias{workflow}
\title{workflow}
\usage{
workflow(files, parameters)
}
\arguments{
\item{files}{character vector of file paths to raw data and info files}

\item{parameters}{S4 object of class WorkflowParameters containing the workflow parameters.}
}
\description{
execute workflow based on specified worklow parameters and a file list.
}
\examples{
\dontrun{
analysis <- workflow(
   list.files(
       system.file(
           'DataSets/FIE-HRMS/BdistachyonEcotypes',
           package = 'metaboData'),
       full.names = TRUE), 
       workflowParameters('FIE_HRMSfingerprinting')
   )}
}
