% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow}
\alias{workflow}
\title{workflow}
\usage{
workflow(files, info, parameters)
}
\arguments{
\item{files}{list of file paths to raw data}

\item{info}{tibble containing sample information}

\item{parameters}{S4 object of class WorkflowParameters containing the workflow parameters.}
}
\description{
execute workflow based on specified worklow parameters and a file list.
}
\examples{
\dontrun{
library(metaboData)
library(binneR)

files <- filePaths('FIE-HRMS','BdistachyonEcotypes') 
info <- runinfo('FIE-HRMS','BdistachyonEcotypes')

wp <- workflowParameters('FIE_HRMSfingerprinting')
wp@processing <- detectParameters(files)
analysis <- workflow(files, info, wp)
}
}
