% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflowParameters.R
\name{workflowParameters}
\alias{workflowParameters}
\title{workflowParameters}
\usage{
workflowParameters(workflow = NULL)
}
\arguments{
\item{workflow}{the workflow analysis to use. NULL prints the available workflows.}
}
\description{
Initiate default workflow parameters for a selected workflow.
}
