% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflowParameters.R
\name{workflowParameters}
\alias{workflowParameters}
\title{workflowParameters}
\usage{
workflowParameters(workflow, files, info, cls = "class", ...)
}
\arguments{
\item{workflow}{the workflow analysis to use. NULL prints the available workflows.}

\item{files}{list of file paths to raw data}

\item{info}{tibble containing sample information}

\item{cls}{info column containing class information to use as default for pre-treatment and modelling}

\item{...}{arguments to pass to binneR::detectParameters}
}
\description{
Initiate default workflow parameters for a selected workflow.
}
\examples{
fp <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes') 
si <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')

wp <- workflowParameters('FIE-HRMS fingerprinting',files,info)
}
